requirements.txt
psutil
requests
Pillow
vt-py   # official VirusTotal client
